+++ tests/Support/TestTrait.php
@@ -60,7 +60,7 @@ trait TestTrait
         return $this->dsn;
     }
 
-    protected function getDriverName(): string
+    protected static function getDriverName(): string
     {
         return 'pgsql';
     }