+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - New #408, #410: Implement `DMLQueryBuilder::upsertReturning()` method (@Tigrov)
 - Enh #412: Reduce binding parameters (@Tigrov)
 - Chg #414: Rename `DMLQueryBuilder::insertWithReturningPks()` to `DMLQueryBuilder::insertReturningPks()` (@Tigrov)
+- Enh #415: Implement `CaseExpressionBuilder` class (@Tigrov)
 
 ## 1.3.0 March 21, 2024